full translate of the website
The greek version of the site is available on the "el" folder, and the
default language is English.
I have added a language change button in the menu.
The way that "#eventsTableThread" I think that does not work, but could
not find why. The translation of the text was made by me, I hope it is
satisfying :)